chore: added FilterNil tests
