--- stringptr.go
@@ -1,5 +1,0 @@
-package slices
-
-func ptr[T any](t T) *T {
-	return &t
-}
